[ANDROSDK-1983] adapt api logic to network layer
