+++ core/src/main/java/org/hisp/dhis/android/network/datavalue/DataValueFields.kt
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2004-2023, University of Oslo
+ *  Copyright (c) 2004-2025, University of Oslo
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -25,23 +25,30 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.core.datavalue.internal
+package org.hisp.dhis.android.network.datavalue
 
-import org.hisp.dhis.android.core.arch.api.executors.internal.APIDownloader
-import org.hisp.dhis.android.core.arch.call.factories.internal.QueryCall
 import org.hisp.dhis.android.core.datavalue.DataValue
-import org.koin.core.annotation.Singleton
+import org.hisp.dhis.android.core.datavalue.DataValueTableInfo.Columns
+import org.hisp.dhis.android.network.common.fields.BaseFields
+import org.hisp.dhis.android.network.common.fields.Fields
 
-@Singleton
-internal class DataValueCall(
-    private val networkHandler: DataValueNetworkHandler,
-    private val handler: DataValueHandler,
-    private val apiDownloader: APIDownloader,
-) : QueryCall<DataValue, DataValueQuery> {
+internal object DataValueFields : BaseFields<DataValue>() {
+    const val ORGANISATION_UNIT = "orgUnit"
+    const val FOLLOW_UP = "followup"
+    const val DELETED = "deleted"
 
-    override suspend fun download(query: DataValueQuery): List<DataValue> {
-        return apiDownloader.downloadListAsCoroutine(handler) {
-            networkHandler.getDataValues(query.bundle)
-        }
-    }
+    val allFields = Fields.from(
+        fh.field(Columns.DATA_ELEMENT),
+        fh.field(Columns.PERIOD),
+        fh.field(ORGANISATION_UNIT),
+        fh.field(Columns.CATEGORY_OPTION_COMBO),
+        fh.field(Columns.ATTRIBUTE_OPTION_COMBO),
+        fh.field(Columns.VALUE),
+        fh.field(Columns.STORED_BY),
+        fh.field(Columns.CREATED),
+        fh.field(Columns.LAST_UPDATED),
+        fh.field(Columns.COMMENT),
+        fh.field(FOLLOW_UP),
+        fh.field(DELETED),
+    )
 }
